+++ gax-grpc/src/test/java/com/google/api/gax/grpc/InstantiatingGrpcChannelProviderTest.java
@@ -67,6 +67,7 @@ import org.threeten.bp.Duration;
 
 @RunWith(JUnit4.class)
 public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportChannelTest {
+
   @Test
   public void testEndpoint() {
     String endpoint = "localhost:8080";
@@ -164,11 +165,8 @@ public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportC
     Duration keepaliveTime = Duration.ofSeconds(1);
     Duration keepaliveTimeout = Duration.ofSeconds(2);
     ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> channelConfigurator =
-        new ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder>() {
-          @Override
-          public ManagedChannelBuilder apply(ManagedChannelBuilder input) {
-            throw new UnsupportedOperationException();
-          }
+        builder -> {
+          throw new UnsupportedOperationException();
         };
     Map<String, ?> directPathServiceConfig = ImmutableMap.of("loadbalancingConfig", "grpclb");
 
@@ -234,10 +232,10 @@ public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportC
     ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> channelConfigurator =
         Mockito.mock(ApiFunction.class);
 
-    ArgumentCaptor<ManagedChannelBuilder> channelBuilderCaptor =
+    ArgumentCaptor<ManagedChannelBuilder<?>> channelBuilderCaptor =
         ArgumentCaptor.forClass(ManagedChannelBuilder.class);
 
-    ManagedChannelBuilder swappedBuilder = Mockito.mock(ManagedChannelBuilder.class);
+    ManagedChannelBuilder<?> swappedBuilder = Mockito.mock(ManagedChannelBuilder.class);
     ManagedChannel fakeChannel = Mockito.mock(ManagedChannel.class);
     Mockito.when(swappedBuilder.build()).thenReturn(fakeChannel);
 
@@ -290,12 +288,10 @@ public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportC
     executor.shutdown();
 
     ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> channelConfigurator =
-        new ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder>() {
-          public ManagedChannelBuilder apply(ManagedChannelBuilder channelBuilder) {
-            // Clients with non-GCE credentials will not attempt DirectPath.
-            assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
-            return channelBuilder;
-          }
+        channelBuilder -> {
+          // Clients with non-GCE credentials will not attempt DirectPath.
+          assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
+          return channelBuilder;
         };
 
     TransportChannelProvider provider =
@@ -320,12 +316,10 @@ public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportC
     executor.shutdown();
 
     ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> channelConfigurator =
-        new ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder>() {
-          public ManagedChannelBuilder apply(ManagedChannelBuilder channelBuilder) {
-            // Clients without setting attemptDirectPath flag will not attempt DirectPath
-            assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
-            return channelBuilder;
-          }
+        channelBuilder -> {
+          // Clients without setting attemptDirectPath flag will not attempt DirectPath
+          assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
+          return channelBuilder;
         };
 
     TransportChannelProvider provider =
@@ -350,12 +344,10 @@ public class InstantiatingGrpcChannelProviderTest extends AbstractMtlsTransportC
     executor.shutdown();
 
     ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> channelConfigurator =
-        new ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder>() {
-          public ManagedChannelBuilder apply(ManagedChannelBuilder channelBuilder) {
-            // Clients without setting attemptDirectPath flag will not attempt DirectPath
-            assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
-            return channelBuilder;
-          }
+        channelBuilder -> {
+          // Clients without setting attemptDirectPath flag will not attempt DirectPath
+          assertThat(channelBuilder instanceof ComputeEngineChannelBuilder).isFalse();
+          return channelBuilder;
         };
 
     TransportChannelProvider provider =